% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validasi}
\alias{validasi}
\title{script for validating data}
\usage{
validasi()
}
\description{
This function allows you go through various steps to cleanse the data
}
\examples{
validate()
}
\keyword{validate,}
\keyword{verified}

